+++ common/src/main/java/net/mehvahdjukaar/polytone/PlatStuff.java
@@ -12,6 +12,7 @@ import net.minecraft.client.particle.ParticleEngine;
 import net.minecraft.client.particle.ParticleProvider;
 import net.minecraft.client.renderer.DimensionSpecialEffects;
 import net.minecraft.client.renderer.RenderType;
+import net.minecraft.client.resources.model.BakedModel;
 import net.minecraft.core.Holder;
 import net.minecraft.core.MappedRegistry;
 import net.minecraft.core.Registry;
@@ -195,4 +196,8 @@ public class PlatStuff {
         throw new AssertionError();
     }
 
+    @ExpectPlatform
+    public static BakedModel getBakedModel(ResourceLocation model) {
+        throw new AssertionError();
+    }
 }