+++ common/src/main/java/net/mehvahdjukaar/polytone/Polytone.java
@@ -8,6 +8,7 @@ import net.mehvahdjukaar.polytone.color.ColorManager;
 import net.mehvahdjukaar.polytone.colormap.ColormapsManager;
 import net.mehvahdjukaar.polytone.dimension.DimensionEffectsManager;
 import net.mehvahdjukaar.polytone.fluid.FluidPropertiesManager;
+import net.mehvahdjukaar.polytone.item.ItemModelOverrideManager;
 import net.mehvahdjukaar.polytone.item.ItemModifiersManager;
 import net.mehvahdjukaar.polytone.lightmap.LightmapsManager;
 import net.mehvahdjukaar.polytone.particle.CustomParticlesManager;
@@ -50,6 +51,7 @@ public class Polytone {
     public static final BlockPropertiesManager BLOCK_MODIFIERS = new BlockPropertiesManager();
     public static final FluidPropertiesManager FLUID_MODIFIERS = new FluidPropertiesManager();
     public static final ItemModifiersManager ITEM_MODIFIERS = new ItemModifiersManager();
+    public static final ItemModelOverrideManager ITEM_MODELS = new ItemModelOverrideManager();
     public static final BiomeEffectsManager BIOME_MODIFIERS = new BiomeEffectsManager();
     public static final ColormapsManager COLORMAPS = new ColormapsManager();
     public static final LightmapsManager LIGHTMAPS = new LightmapsManager();
@@ -73,7 +75,7 @@ public class Polytone {
     public static void init( boolean devEnv, boolean forge) {
         PlatStuff.addClientReloadListener(() -> new CompoundReloader(
                         SOUND_TYPES, BIOME_ID_MAPPERS, COLORMAPS, CUSTOM_PARTICLES, COLORS,
-                        BLOCK_SET, BLOCK_MODIFIERS, FLUID_MODIFIERS, ITEM_MODIFIERS,
+                        BLOCK_SET, BLOCK_MODIFIERS, FLUID_MODIFIERS, ITEM_MODELS, ITEM_MODIFIERS,
                         BIOME_MODIFIERS, VARIANT_TEXTURES, LIGHTMAPS, DIMENSION_MODIFIERS,
                         PARTICLE_MODIFIERS, SLOTIFY, OVERLAY_MODIFIERS,
                         CREATIVE_TABS_MODIFIERS),