+++ common/src/main/java/net/mehvahdjukaar/polytone/dimension/DimensionEffectsModifier.java
@@ -1,5 +1,6 @@
 package net.mehvahdjukaar.polytone.dimension;
 
+import com.mojang.datafixers.util.Either;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.Decoder;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
@@ -18,7 +19,7 @@ import java.util.Optional;
 import java.util.Set;
 
 
-public record DimensionEffectsModifier(Optional<Float> cloudLevel,
+public record DimensionEffectsModifier(Optional<Either<Float, ColormapExpressionProvider>> cloudLevel,
                                        Optional<Boolean> hasGround,
                                        Optional<DimensionSpecialEffects.SkyType> skyType,
                                        Optional<Boolean> forceBrightLightmap,
@@ -27,7 +28,6 @@ public record DimensionEffectsModifier(Optional<Float> cloudLevel,
                                        Optional<IColorGetter> skyColor,
                                        boolean noWeatherFogDarken,
                                        boolean noWeatherSkyDarken,
-                                       Optional<ColormapExpressionProvider> cloudHeight,
                                        Optional<Lightmap> lightmap,
                                        Set<ResourceLocation> explicitTargets) implements ITargetProvider {
 
@@ -36,7 +36,7 @@ public record DimensionEffectsModifier(Optional<Float> cloudLevel,
 
     public static final Decoder<DimensionEffectsModifier> CODEC = RecordCodecBuilder.create(instance ->
             instance.group(
-                    Codec.FLOAT.optionalFieldOf("cloud_level").forGetter(DimensionEffectsModifier::cloudLevel),
+                    Codec.either(Codec.FLOAT, ColormapExpressionProvider.CODEC).optionalFieldOf("cloud_level").forGetter(DimensionEffectsModifier::cloudLevel),
                     Codec.BOOL.optionalFieldOf("has_ground").forGetter(DimensionEffectsModifier::hasGround),
                     SKY_TYPE_CODEC.optionalFieldOf("sky_type").forGetter(DimensionEffectsModifier::skyType),
                     Codec.BOOL.optionalFieldOf("force_bright_lightmap").forGetter(DimensionEffectsModifier::forceBrightLightmap),
@@ -50,15 +50,15 @@ public record DimensionEffectsModifier(Optional<Float> cloudLevel,
             ).apply(instance, DimensionEffectsModifier::new));
 
     public static DimensionEffectsModifier ofFogColor(Colormap colormap) {
-        return new DimensionEffectsModifier(java.util.Optional.empty(), java.util.Optional.empty(), java.util.Optional.empty(), java.util.Optional.empty(),
-                java.util.Optional.empty(), Optional.of(colormap), java.util.Optional.empty(),
-                false, false, java.util.Optional.empty(), Set.of());
+        return new DimensionEffectsModifier(Optional.empty(), Optional.empty(), Optional.empty(), Optional.empty(),
+                Optional.empty(), Optional.of(colormap), Optional.empty(),
+                false, false , Optional.empty(), Set.of());
     }
 
     public static DimensionEffectsModifier ofSkyColor(Colormap colormap) {
-        return new DimensionEffectsModifier(java.util.Optional.empty(), java.util.Optional.empty(), java.util.Optional.empty(), java.util.Optional.empty(),
-                java.util.Optional.empty(), java.util.Optional.empty(), Optional.of(colormap),
-                false, false, java.util.Optional.empty(), Set.of());
+        return new DimensionEffectsModifier(Optional.empty(), Optional.empty(), Optional.empty(), Optional.empty(),
+                Optional.empty(), Optional.empty(), Optional.of(colormap),
+                false, false, Optional.empty(), Set.of());
     }
 
 
@@ -88,10 +88,10 @@ public record DimensionEffectsModifier(Optional<Float> cloudLevel,
 
     public DimensionEffectsModifier applyInplace(ResourceLocation dimensionId) {
         DimensionSpecialEffects effects = PlatStuff.getDimensionEffects(dimensionId);
-        Optional<Float> oldCloud = Optional.empty();
-        if (this.cloudLevel.isPresent()) {
-            oldCloud = Optional.of(effects.cloudLevel);
-            effects.cloudLevel = this.cloudLevel.get();
+        Optional<Either<Float, ColormapExpressionProvider>> oldCloud = Optional.empty();
+        if (this.cloudLevel.isPresent() && this.cloudLevel.get().left().isPresent()) {
+            oldCloud = Optional.of(Either.left(effects.cloudLevel));
+            effects.cloudLevel = this.cloudLevel.get().left().get();
         }
         Optional<Boolean> oldGround = Optional.empty();
         if (this.hasGround.isPresent()) {