+++ common/src/main/java/net/mehvahdjukaar/polytone/mixins/LevelRendererMixin.java
@@ -1,0 +1,25 @@
+package net.mehvahdjukaar.polytone.mixins;
+
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
+import net.mehvahdjukaar.polytone.Polytone;
+import net.minecraft.client.multiplayer.ClientLevel;
+import net.minecraft.client.renderer.LevelRenderer;
+import org.jetbrains.annotations.Nullable;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(LevelRenderer.class)
+public class LevelRendererMixin {
+
+    @Shadow
+    @Nullable
+    private ClientLevel level;
+
+    @ModifyExpressionValue(method = "renderClouds", at = @At(value = "INVOKE",
+            target = "Lnet/minecraft/client/renderer/DimensionSpecialEffects;getCloudHeight()F"))
+    private float polytone$modifyCloudHeight(float original) {
+        Float f = Polytone.DIMENSION_MODIFIERS.modifyCloudHeight(this.level);
+        return f != null ? f : original;
+    }
+}