+++ fabric/src/main/java/net/mehvahdjukaar/polytone/mixins/fabric/ModelManagerAccessor.java
@@ -1,0 +1,17 @@
+package net.mehvahdjukaar.polytone.mixins.fabric;
+
+import net.minecraft.client.resources.model.BakedModel;
+import net.minecraft.client.resources.model.ModelManager;
+import net.minecraft.client.resources.model.ModelResourceLocation;
+import net.minecraft.resources.ResourceLocation;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+import java.util.Map;
+
+@Mixin(ModelManager.class)
+public interface ModelManagerAccessor {
+
+    @Accessor("bakedRegistry")
+    Map<ModelResourceLocation, BakedModel> getBakedRegistry();
+}